refactor(table): reorganize field-related imports and introduce new field components for improved structure
