+++ components/table/fields/field-icon.tsx
@@ -1,6 +1,6 @@
 import { FieldType } from "@/lib/fields/const"
 
-import { makeHeaderIcons } from "./views/grid/fields/header-icons"
+import { makeHeaderIcons } from "./header-icons"
 
 const icons = makeHeaderIcons(18)
