+++ components/table/fields/property/number/number-property-editor.tsx
