+++ components/table/fields/property/text/text-property-editor.tsx
